+++ src/App.jsx
@@ -2,7 +2,8 @@ import { BrowserRouter, Route, Routes } from "react-router-dom"
 import { LoginRegister } from "./pages/LoginRegister";
 import { LandingPage } from "./pages/LandingPage";
 import { HomeUser } from "./pages/HomeUser";
-import { VistaAdministrador } from "./pages/VistaAdministrador";
+import { UsersAdmin } from "./pages/UsersAdmin";
+import { SensoresAdmin } from "./pages/SensoresAdmin";
 
 function App() {
 
@@ -13,7 +14,8 @@ function App() {
           <Route path="/" element={<LandingPage/>}/>
           <Route path="/login" element={<LoginRegister/>}/>
           <Route path="/home" element={<HomeUser/>}/>
-          <Route path="/Admin" element={<VistaAdministrador/>}/>
+          <Route path="/users" element={<UsersAdmin/>}/>
+          <Route path="/sensores" element={<SensoresAdmin/>}/>
         </Routes>
       </BrowserRouter>
     </>
