+++ src/assets/Styles/AdminNavBar.css
@@ -1,12 +1,12 @@
 @import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600;700;900&display=swap');
 
-.navbarUser{
+.NavBarAdmin{
     width: 15vh;
     height: 100vh;
     background-color: #FFFFFF;
     overflow: hidden;
 }
-.nav{
+.Nav{
     display: flex;
     justify-content: center;
 }
