+++ src/components/TableSensor.jsx
@@ -1,0 +1,11 @@
+export const TableSensor = () =>{
+    return(
+        <>
+            <h2 className="users">Lista de sensores</h2>
+            <div className="container-crud">
+                <table className="table-sensores">
+                </table>
+            </div>
+        </>
+    )
+}
