+++ src/components/UserNavBar.jsx
@@ -8,23 +8,22 @@ import '../assets/Styles/userNavBar.css';
 const UserNavBar = () => {
     return (
         <>
-            
-                <div className="row">
-                    <div className="col nav">
-                        <a href=""><img src={Logo} alt="Logo" id='LogoAQA' /></a>
-                    </div>
+            <div className="row">
+                <div className="col nav">
+                    <a href=""><img src={Logo} alt="Logo" id='LogoAQA' /></a>
                 </div>
-                <div className="row">
-                    <div className="col optionsUser nav">
-                        <a href=""><img src={Home} alt="Home" className='options' /></a>
-                        <a href=""><img src={Settings} alt="Settings" className='options' /></a>
-                    </div>
+            </div>
+            <div className="row">
+                <div className="col optionsUser nav">
+                    <a href=""><img src={Home} alt="Home" className='options' /></a>
+                    <a href=""><img src={Settings} alt="Settings" className='options' /></a>
                 </div>
-                <div className="row">
-                    <div className="col nav">
-                        <a href=""><img src={logout} alt="" id='Logout' /></a>
-                    </div>
+            </div>
+            <div className="row">
+                <div className="col nav">
+                    <a href=""><img src={logout} alt="" id='Logout' /></a>
                 </div>
+            </div>
         </>
     )
 }
