+++ src/pages/CambioContraseña.jsx
@@ -1,0 +1,6 @@
+export const CambioContraseña = () =>{
+    return(
+        <>
+        </>
+    )
+}
